Separate test files for easier maintenance.
